Add internal family name and replace with do bet
